docs/tutorials/cat_dogs/dog_cat_backend/requirements-base.txt
gunicorn==22.0.0
label-studio-ml @ git+https://github.com/HumanSignal/label-studio-ml-backend.git
transformers==4.44.2
torch>=2.0.0
Pillow==12.0.0
numpy==2.2.6
scikit-image==0.25.2
lime==0.2.0.1
requests==2.31.0
python-dotenv==1.1.1
tqdm==4.67.1
